Add runtime dependencies' versions
This commit adds new entries in the versions.txt for indicating the
versions of the OBS packages in the stable repository. These are
added to the spec/dsc files in the "Requires" section and in order
to support the upgrade path when updating the runtime.

Fixes #21

Update runtime and shim to 3.0.0-alpha.4

- Update runtime and shim to 3.0.0-alplha.4
- Update proxy to 3.0.0-alpha.3
- Move dependencies to the metapackage.

Signed-off-by: Erick Cardona <erick.cardona.ruiz@intel.com>