linux-container: Simplify build of linux-continer with OBS
This commit will bring a faster way to build linux kernels for
clear containers using the source _service that OBS provides.

With this commit the requeriment to install debtools in the developer
enviroment is gone, also a bunch of overhead with the upload of ~88Mb
of the linux kernel source, adding speed to the build of the package.

More details:
https://en.opensuse.org/openSUSE:Build_Service_Concept_SourceService

Fixes #72

Signed-off-by: Geronimo Orozco <geronimo.orozco@intel.com>