runtime: Remove staging code
This commit removes the option STAGING. This option would allow
to build the runtime wothout dependencies. Now, this option does not
seem to help anymore since we have full control over the versions and
revisions statically. By default, the runtime will be built with the
versions specified in the versions.txt file.

Fixes #106

Signed-off-by: Erick Cardona <erick.cardona.ruiz@intel.com>