ksm: Introduce ksm-throttler package
This commit adds the template files and update script for the
ksm-throttler package.
Since the upstream repository
(https://github.com/kata-containers/ksm-throttler) does not have a
release version yet, the version string for this package has been
initialized with "0.0.1". The version string also contains the sha
of the latest commit.

Fixes: #198

Signed-off-by: Erick Cardona <erick.cardona.ruiz@intel.com>