kernel: auto generate kernel config before update
Right now, kernel packages are generated from the kernel config that
lives in clearcontainers/packaging repository. It does not get upgraded
after a kernel version update.
This commit add the capability to generate a new config file, based on
the existing one, from the kernel tree of the version that is intended
to be packaged. This new config file is used to build newer kernel
packages.

Fixes #207

Signed-off-by: Erick Cardona <erick.cardona.ruiz@intel.com>