kernel: Create tool to check for new LTS kernel releases
We need to be notified when new LTS kernel releases are available to ensure we
update the guest kernel in a timely fashion.
This commit adds a documentation and a script which can check the latest kernel info
and notify via email if a new LTS version is released.

Fixes #246

Signed-off-by: Erick Cardona <erick.cardona.ruiz@intel.com>